Componentize txn detail switch
